initial shoping list in db

+++ lib/movie/movie.dart
@@ -11,7 +11,7 @@ class Movie {
   Movie.formJson(Map <String, dynamic> parsedJson){
     this.id = parsedJson['id'];
     this.title = parsedJson['title'];
-    this.voteAverage = parsedJson['vote_average'];
+    this.voteAverage = parsedJson['vote_average'].toDouble();
     this.releaseDate = parsedJson['release_date'];
     this.overview = parsedJson['overview'];
     this.posterPath = parsedJson['poster_path'];
